services/api-gateway/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
pydantic==2.5.3
python-jose[cryptography]==3.3.0
redis==5.0.1
